% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_get_descriptives.R
\name{ee_get_descriptives}
\alias{ee_get_descriptives}
\title{Compute Descriptive Statistics for Emotion Expression Time Series Data}
\usage{
ee_get_descriptives(dat)
}
\arguments{
\item{dat}{A dataframe containing the time series data. The dataframe must
include a "time" column and one or more variables representing different
emotions or signals to analyze.}
}
\value{
A dataframe (tibble) containing the descriptive statistics for each
emotion variable: number of missing values, number of complete cases,
percentage of complete cases, mean, median, standard deviation, 5th percentile,
95th percentile, minimum, and maximum values.
}
\description{
This function takes a dataframe with time series data and computes descriptive
statistics for each EE variable present in the dataset. It reshapes the data,
calculates the number of missing and complete cases, the percentage of complete cases,
and standard descriptive statistics (mean, median, standard deviation, 5th percentile,
95th percentile, minimum, and maximum) for each emotion.
}
\examples{
\dontrun{
# Assumes `ee_read` function is used to read time series data
dat <- ee_read("inst/extdata/pt1.txt")
descriptives <- ee_get_descriptives(dat)
print(descriptives)
}

}
\author{
Nate Hall
}
