% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_read.R
\name{ee_read}
\alias{ee_read}
\title{Read and Tidy Noldus FaceReader Data File}
\usage{
ee_read(fpath, id = NULL)
}
\arguments{
\item{fpath}{Path to the Noldus FaceReader output file (.txt).}

\item{id}{(Optional) Participant ID. If provided, will be added as a column.}
}
\value{
A tibble (data frame) containing:
\itemize{
 \item \code{id}: Participant ID (if provided).
 \item \code{time}: Time point in seconds.
 \item \code{neutral}, \code{happy}, \code{sad}, \code{angry},
       \code{surprised}, \code{scared}, \code{disgusted}:
       Emotion expression intensities (likely ranging from 0 to 1).
 \item \code{valence}: Valence of emotion expression (positive/negative).
 \item \code{arousal}: Arousal level of emotion expression.
}
}
\description{
Reads in emotion expression data from Noldus FaceReader
V 8.0 output files (.txt format) and performs basic setup of data object.
}
\examples{
# Read data from a file for test participant 1
my_data <- ee_read("data/pt1.txt", id = 1)

}
