% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_gimme_graph_properties.R
\name{extract_gimme_graph_properties}
\alias{extract_gimme_graph_properties}
\title{Extract Graph Properties from GIMME Analysis}
\usage{
extract_gimme_graph_properties(
  betas,
  lagged = FALSE,
  outliers = NULL,
  edge_values_from = "beta",
  silent = TRUE,
  keep_igraph = FALSE
)
}
\arguments{
\item{betas}{A dataframe containing the beta coefficients from a GIMME analysis.
This dataframe should include, at a minimum, columns for the left-hand side
(lhs), the right-hand side (rhs), and the beta values.}

\item{lagged}{Logical indicating whether to create a graph from lagged paths
only (`TRUE`) or contemporaneous paths only (`FALSE`). Defaults to `FALSE`.}

\item{outliers}{An optional list specifying parameters for outlier adjustment
within the graph properties analysis. The list can include methods for outlier
handling and limits for identification, among other settings.}

\item{edge_values_from}{Character string specifying the source of edge values.
Options are "beta" for beta coefficients or "z" for z-scores. Default is "beta".}

\item{silent}{Logical indicating whether to suppress messages during function
execution. Defaults to `TRUE`.}

\item{keep_igraph}{Logical indicating whether to include the raw igraph objects
in the function's output. Defaults to `FALSE`.}
}
\value{
A list containing the extracted graph properties. If `keep_igraph` is
`TRUE`, the list includes three elements: `edges`, which contains the dataframe
of edges with potential outlier adjustments; `graph_metrics`, a list of dataframes
with graph metrics for each file; and `graphs`, a list of igraph objects for
each file. If `keep_igraph` is `FALSE`, the `graphs` element is omitted.
}
\description{
Extracts and computes graph properties from GIMME beta coefficients, allowing
the selection between contemporaneous and cross-lagged edges. The function can
handle outlier adjustment and offers options to retain the raw igraph objects
for further custom analyses. This functionality facilitates in-depth exploration
of the network characteristics derived from GIMME outputs, providing insights
into the dynamics of the system under study.
}
\examples{
# Assuming `betas` is a dataframe of beta coefficients from GIMME analysis
graph_props <- extract_gimme_graph_properties(betas, lagged = FALSE, keep_igraph = TRUE)

}
\author{
Nate Hall
}
