% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_na_edge_rows.R
\name{get_na_edge_rows}
\alias{get_na_edge_rows}
\title{Identify NA Rows at the Edges of Time Series Columns}
\usage{
get_na_edge_rows(
  dat,
  cols = c("neutral", "happy", "sad", "angry", "surprised", "scared", "disgusted",
    "valence", "arousal")
)
}
\arguments{
\item{dat}{A data frame containing the time series data.}

\item{cols}{A character vector specifying the names of the columns in `dat` to check
for NA rows at the edges. Typically, this excludes identifier and time columns.}
}
\value{
A list where each element corresponds to a column specified in `cols`. Each
list element contains a numeric vector of row numbers that are NA at the beginning
or end of the time series for that column. If a column has no NAs at its edges, it
will not appear in the list.
}
\description{
This function iterates over specified columns in a data frame, identifying rows
that contain NA values at the beginning and end of the time series. It's particularly
useful for preprocessing steps before interpolation, allowing users to handle edge
cases where NA values may not be suitable for interpolation without a valid anchor.
}
\examples{
dat <- data.frame(id = 1:10, time = 1:10, measure1 = c(NA, NA, 3:8, NA, NA), measure2 = c(1:4, NA, NA, NA, 7:10))
interp_cols <- setdiff(colnames(dat), c("id", "time"))
na_edge_rows <- get_na_edge_rows(dat, interp_cols)
# Check the output to see the rows with NAs at the edges for 'measure1' and 'measure2'

}
