% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_outliers_bygroup_oneside.R
\name{handle_outliers_bygroup_oneside}
\alias{handle_outliers_bygroup_oneside}
\title{One-sided Outlier Handling by Group}
\usage{
handle_outliers_bygroup_oneside(
  df,
  columns,
  side = "lo",
  method = "exclude",
  limit_class = "prob",
  limit = 0.025,
  groupBy = NULL
)
}
\arguments{
\item{df}{A dataframe containing the data to be processed.}

\item{columns}{A character vector specifying the columns within `df` to be
checked for outliers.}

\item{side}{A character string indicating the side of the outlier. Valid
options are "lo" for lower side or "hi" for higher side.}

\item{method}{A character string specifying the method to handle outliers.
Valid options are "exclude" (default) or "winsorize".}

\item{limit_class}{A character string indicating the type of limit to use.
Valid options are "prob" for probabilistic limits based on quantiles or
"fixed" for fixed value limits. Default is "prob".}

\item{limit}{A numeric value specifying the limit for outlier detection.
When `limit_class` is "prob", `limit` represents a quantile (e.g., 0.025 for
lower 2.5%); when `limit_class` is "fixed", it represents the actual value
threshold.}

\item{groupBy}{An optional character vector specifying columns to group by
before performing outlier handling. If NULL (default), outlier handling is
applied column-wise across the entire dataset.}
}
\value{
A dataframe with outliers handled according to the specified method
and parameters. If the method is "exclude", outliers are set to NA. If the
method is "winsorize", outlier values are adjusted.
}
\description{
Handles outliers in specified columns of a dataframe on one side (lower or
higher) based on a given method (exclude or winsorize), limit classification
(probability or fixed value), and an optional grouping. This function can
set outliers to NA (for exclusion) or adjust their values (for winsorization).
}
\examples{
# Assuming `df` is your dataframe, and you want to handle lower outliers
# in the "weight" column by exclusion, based on the bottom 2.5\% quantile
cleaned_df <- handle_outliers_bygroup_oneside(df, columns = "weight",
                                              side = "lo", method = "exclude",
                                              limit_class = "prob", limit = .025)

}
\author{
Nate Hall
}
