% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ee_timeseries.R
\name{plot_ee_timeseries}
\alias{plot_ee_timeseries}
\title{Plot Emotion Expression Time Series}
\usage{
plot_ee_timeseries(
  sub_ts,
  include_raw = TRUE,
  id = NULL,
  desc = NULL,
  plot_type = "combined"
)
}
\arguments{
\item{sub_ts}{A data frame containing the subject's emotion time series data.}

\item{include_raw}{Logical, indicating whether to include raw emotion data in the plot.
Default is `TRUE`.}

\item{id}{Optional; the subject's identifier. If provided, it will be used in the plot title.}

\item{desc}{Optional; a data frame returned by `ee_get_descriptives` containing descriptive
statistics of the subject's data. If provided, the percentage of readable frames
is included in the plot subtitle.}

\item{plot_type}{Character string specifying the type of plot: "combined" (default),
"discrete", or "circumplex". Determines whether to plot all data combined,
just the discrete emotions, or just the circumplex model (valence and arousal).}
}
\value{
A `ggplot` object representing the emotion time series plot as specified
        by the `plot_type` parameter.
}
\description{
This function visualizes the time series of facial emotion expression data from FaceReader. It supports the plotting
of raw emotion intensities over time and can distinguish between discrete emotions
and circumplex model (valence and arousal). Users can specify the type of plot desired:
combined (default), discrete emotions, or circumplex model only.
}
\examples{
plot_ee_timeseries(sub_ts = my_data, id = "Subject1", desc = descriptives,
                   plot_type = "combined")

}
\author{
Nate Hall
}
