% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_interp.R
\name{ts_interp}
\alias{ts_interp}
\title{Interpolate Missing Values}
\usage{
ts_interp(dat, maxgap = 30, interp_option = "stine", range_restrict = TRUE)
}
\arguments{
\item{dat}{A data frame containing time series data, with the first column
assumed to be a time/index variable.}

\item{maxgap}{Maximum gap size (number of consecutive NAs) allowed
for interpolation. Defaults to 30 (for FaceReader data collected at 30 Hz, equivalent to smoothing over 1 second).}

\item{interp_option}{The interpolation option to use from the `imputeTS`
package. Can take values, 'linear', 'spline', 'stine'.
Defaults to "stine".}

\item{range_restrict}{Logical. When TRUE (Default), perform range restriction on EE timeseries}
}
\value{
A data frame with interpolated values and (if requested) adjusted ranges. If range_restrict is TRUE, the range-adjusted columns will be added with _adjusted appended to the column names.
}
\description{
This function performs interpolation on missing values within time series
data, followed by range adjustments if requested to ensure EE columns remain within
their valid ranges (0-1 for all EEs other than valence, which is adjusted to [-1, 1].
}
\examples{
dat <- ee.dyn::ee_read("dev/data/pt1.txt", id = 1)
result <- ts_interp(dat)

}
\author{
Nate Hall
}
